+++ NEWS.md
@@ -1,4 +1,4 @@
-# mirai 0.8.3.9019 (development)
+# mirai 0.8.3.9020 (development)
 
 * New `mirai_version()` utility provides the package version, and versions of its depdendencies (`nanonext`, NNG and mbed TLS libraries) in a single string.
 * The deferred evaluation pipe `%>>%` gains the following enhancements:
@@ -6,7 +6,7 @@
   + expressions may be tested using `unresolved()` in the same way as a 'mirai'.
   + allows for general use in all contexts, including within functions.
 * Improved error messages for top level evaluation errors in a 'mirai'.
-* Requires nanonext >= [0.8.2.9016].
+* Requires nanonext >= [0.8.2.9017].
 * Internal performance enhancements.
 
 # mirai 0.8.3